requirements.txt
ib_insync
pandas
numpy
backtesting
xgboost
scikit-learn
python-dotenv
requests
tabulate
statsmodels
matplotlib
seaborn
# TA-Lib
yfinance
hmmlearn
streamlit
plotly